DB-Integrations

+++ API-GruasUCAB-main/API-GruasUCAB.Policy/Infrastructure/DTOs/PolicyQueries/PolicyDTO.cs
@@ -5,10 +5,10 @@ namespace API_GruasUCAB.Policy.Infrastructure.DTOs.PolicyQueries
           public Guid PolicyId { get; set; }
           public string Number { get; set; } = string.Empty;
           public string Name { get; set; } = string.Empty;
-          public float CoverageAmount { get; set; }
+          public decimal CoverageAmount { get; set; }
           public int CoverageKm { get; set; }
-          public float BaseAmount { get; set; }
-          public float PriceKm { get; set; }
+          public decimal BaseAmount { get; set; }
+          public decimal PriceKm { get; set; }
           public string IssueDate { get; set; } = string.Empty;
           public string ExpirationDate { get; set; } = string.Empty;
           public Guid ClientId { get; set; }
